requirements.txt
flask
transformers
numpy
pandas
requests
beautifulsoup4
langchain 
openai 
sentence-transformers 
faiss-cpu
uvicorn
ctransformers
python-box
streamlit
langchain-ollama
ollama
langchain-community
fastapi
langchain_huggingface
chardet
python-docx